requirements.txt
streamlit
requests
pillow
python-dotenv
openai
groq
ollama